requirements/test.txt
beautifulsoup4==4.3.2
factory_boy==2.4.1
mock==1.0.1
requests-mock==0.6.0
nose==1.3.4
unittest2==0.8.0
WebTest==2.0.18
